created a way for all pages to acces the formUtils functions
