+++ src/routes/login/+page.svelte
@@ -1,8 +1,8 @@
 <script>
+    import { areFieldsFilled, baseInputClasses } from "$lib/formUtils";
     import { user } from "$lib/store";
 
     let userFound = true;
-    let fieldsFilledBool = true;
     let fields = {
         user: null,
         password: null
@@ -25,23 +25,20 @@
         }
     }
 
-    function areFieldsFilled(object) {
-        const values = Object.values(object);
-        const index = values.findIndex(value => value === "");
-        if (index === -1) {
-            return true;
-        }
-        return false;
-    }
+    
 </script>
 
 <div class="flex flex-row items-center justify-center min-h-screen bg-gray-100">
     <section class="bg-white shadow-md rounded-lg p-8 w-full max-w-sm">
         <h1 class="text-2xl font-semibold text-gray-800 mb-6 text-center">Inloggen</h1>
         
-        <input id="user" placeholder="Gebruikersnaam of email" class="w-full px-4 py-2 mb-4 text-gray-700 rounded-lg focus:ring-2 focus:ring-blue-500 focus:outline-none border {fields.user === "" ? 'border-red-500' : 'border-gray-300'}"/>
+        <input id="user" placeholder="Gebruikersnaam of email"
+            class="{baseInputClasses} {fields.user === "" ? 'border-red-500' : 'border-gray-300'}"
+        />
         
-        <input id="password" type="password" placeholder="Wachtwoord" class="w-full px-4 py-2 mb-4 text-gray-700 rounded-lg focus:ring-2 focus:ring-blue-500 focus:outline-none border {fields.password === "" ? 'border-red-500' : 'border-gray-300'}"/>
+        <input id="password" type="password" placeholder="Wachtwoord"
+            class="{baseInputClasses} {fields.password === "" ? 'border-red-500' : 'border-gray-300'}"
+        />
 
         {#if !areFieldsFilled(fields)}
             <p class="text-sm text-red-500 mb-4">Vul alles in</p>
